Relocate Proc.
